requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
fastapi==0.95.1
uvicorn==0.21.1
geopy==2.3.0
joblib==1.2.0
python-dotenv==1.0.0
pytest==7.3.1
black==23.3.0